requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.2.0
razorpay==1.4.2
reportlab==4.0.7
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0


# Media storage (S3 via django-storages)
django-storages==1.14.4
boto3==1.34.69
PyPDF2==3.0.1
